Image_models/Domain_adaptation_models/CBD/requirements.txt
numpy==2.1.3
matplotlib==3.9.2
pillow==11.0.0
torch==2.5.1